Updated Fireball (Issue # 28) and Monster Classes
Changed Monster class to abstract and implemented child class functions in ball, bug, frog, glider, tank, and walker classes
