+++ DungeonCrawl/src/model/Moveable/Tank.java
@@ -5,6 +5,9 @@
  */
 package model.Moveable;
 
+import java.awt.Graphics;
+import java.awt.geom.Rectangle2D;
+
 /**
  *
  * @author russe_000
@@ -12,5 +15,30 @@ package model.Moveable;
 public class Tank extends Monster{
     public Tank(float x, float y) {
         super(x, y);
+    }
+
+    @Override
+    public void render(Graphics g) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public boolean isAlive() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public Rectangle2D.Double getCollisionBox() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void update() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void findCollision() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 }
